Add default overwrite 2FA for testing

+++ src/main/java/com/othr/rentopia/controller/UserController.java
@@ -167,22 +167,26 @@ public class UserController {
 
         //validate token and code
         TwoFAToken token = twoFATokenService.getTokenByValue(tokenValue);
-        if(token == null){
-            authResponse.setMessage("no_token");
-            return new ResponseEntity<>(authResponse, HttpStatus.OK);
-        }
 
-        if(token.getExpiration().isBefore(LocalDateTime.now())){
-            authResponse.setMessage("token_expired");
-            return new ResponseEntity<>(authResponse, HttpStatus.OK);
-        }
+        // DEBUG
+        if (! "123456".equals(authCode)) {
+            if(token == null){
+                authResponse.setMessage("no_token");
+                return new ResponseEntity<>(authResponse, HttpStatus.OK);
+            }
 
-        if(!token.getAuthCode().equals(authCode)){
-            authResponse.setMessage("token_invalid");
-            return new ResponseEntity<>(authResponse, HttpStatus.OK);
-        }
+            if(token.getExpiration().isBefore(LocalDateTime.now())){
+                authResponse.setMessage("token_expired");
+                return new ResponseEntity<>(authResponse, HttpStatus.OK);
+            }
 
-        twoFATokenService.removeTokenById(token.getId());
+            if(!token.getAuthCode().equals(authCode)){
+                authResponse.setMessage("token_invalid");
+                return new ResponseEntity<>(authResponse, HttpStatus.OK);
+            }
+
+            twoFATokenService.removeTokenById(token.getId());
+        }
 
         //return jwt
         Account user = accountService.getAccountWithPassword(token.getUserEmail());
